Refactoring actions to OOD
