+++ melissa/actions/repeat/main.py
@@ -1,0 +1,9 @@
+# Melissa
+WORDS = {'repeat_text': {'groups': ['repeat', 'say']}}
+
+
+class Repeat():
+
+    def repeat_text(self, text):
+        text = text.split(' ', 1)[1]
+        return text
